backend/services/insight/pyproject.toml
[project]
name = "insight"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "aio-pika>=9.4.0",
    "redis[hiredis]>=5.0.0",       # idempotency guard + cache
    "httpx>=0.27.0",               # LLM API calls (Anthropic/OpenAI)
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
    "structlog>=24.2.0",
    # RAG deps — installed now, wired in phase 2
    "google-genai>=1.0.0",         # Gemini streaming LLM (new official SDK)
    "pgvector>=0.3.0",             # postgres vector extension client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/insight"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]